Move security related classes into security package
